# perf: Logger 高性能优化与全面测试覆盖提升