--- a/.claude/settings.local.json
+++ b/.claude/settings.local.json
@@ -15,7 +15,9 @@
       "Bash(gh pr view:*)",
       "Bash(go list:*)",
       "Bash(git commit:*)",
-      "Bash(timeout 30s go test -cover -coverprofile=coverage-improved.out ./...)"
+      "Bash(timeout 30s go test -cover -coverprofile=coverage-improved.out ./...)",
+      "Bash(timeout 30s go test -cover -coverprofile=coverage.out ./...)",
+      "Bash(timeout 30s go test -cover -coverprofile=coverage-entry.out ./...)"
     ],
     "deny": [],
     "ask": []