## Summary

本 PR 完成了 Logger 库的高性能优化和全面测试覆盖提升，实现了显著的性能改进：

### 🚀 核心性能优化

**Logger 性能提升**：
- **最大性能模式**: 2.6x 速度提升，5x 内存减少 (438 ns/op vs 1138 ns/op)
- **禁用 Caller**: 2.4x 速度提升，4.5x 内存减少
- **新增配置 API**: `EnableCaller()` 和 `EnableTrace()` 方法

**Entry 系统优化**：
- 优化 Reset 方法：批量字段赋值提升缓存友好性
- 新增内联函数：`getEntry()` 和 `putEntry()` 性能优化
- 深度内存分析：200字节结构体布局优化

### 📊 测试覆盖提升

- **覆盖率**: 从 92.9% 提升至 94.2%
- **函数覆盖**: 125/138 函数达到 100% 覆盖率（排除 Fatal/Panic）
- **新增测试**: Entry 全面测试、性能基准测试、边界条件测试

### ⚡ 性能基准对比

```
标准模式:     1138 ns/op,  360 B/op,  9 allocs/op
最大性能:      438 ns/op,   72 B/op,  4 allocs/op
禁用Caller:    470 ns/op,   80 B/op,  5 allocs/op  
禁用Trace:    1048 ns/op,  344 B/op,  8 allocs/op
```

### 🔧 技术实现

**条件性功能控制**：
- 智能分支避免昂贵的调用者信息获取
- 可配置的跟踪 ID 生成
- 零拷贝前缀后缀处理

**高性能缓存池**：
- 专用 Entry 缓存池替代全局对象池
- 内联函数减少函数调用开销
- 批量字段设置提高内存访问效率

### 🔄 向后兼容

- 完全保持现有 API 兼容性
- 默认启用所有功能，行为不变
- 新增性能配置为可选功能

## Test plan

- [x] 运行所有现有测试确保功能正确性
- [x] 性能基准测试验证优化效果
- [x] 多构建标签测试 (debug/release/discard)
- [x] 并发安全性测试
- [x] 内存泄漏检查

🤖 Generated with [Claude Code](https://claude.ai/code)